VIC-4784 VIC-4779 Misc PR Demo requests (#2677)
+ When resetting the robot by entering the pairing screen, it no
  longer does the big greeting
+ When the robot is picked up, it should play the initial picked up
  animation then the react to on side animations
+ Exploring should bump objects of any size, not just small ones
  (with some probability)
+ Add "go to sleep"
+ Blackjack should always play "bust" animation

(cherry picked from commit 969fc2828b0e5d3ccb9e13b24fa39f351dcae62c)

+++ engine/aiComponent/behaviorComponent/behaviorComponentMessageHandler.cpp
@@ -336,7 +336,7 @@ void BehaviorComponentMessageHandler::SubscribeToWebViz(BehaviorExternalInterfac
         // also send them the list of behaviorIDs that can be created
         Json::Value allBehaviors;
         auto& list = allBehaviors["list"];
-        for( uint8_t i=0; i<BehaviorTypesWrapper::GetBehaviorIDNumEntries(); ++i ) {
+        for( uint16_t i=0; i<BehaviorTypesWrapper::GetBehaviorIDNumEntries(); ++i ) {
           list.append( EnumToString( static_cast<BehaviorID>(i) ) );
         }
         sendToClient( list );
